コンフリクトの修正で大石君のぶんが混じってしまっていたので、一度戻して再度コンフリクトを修正した。
