--- BattaJump/Assets/Script/Audio/AudioSettingUiController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 694efec2fac792d48a0c8d296fafb041
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
