--- BattaJump/Assets/Script/Audio/Audios.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 0709f239f84ab2b4ba686c495ea3fee6
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
