+++ BattaJump/Assets/Script/Audio/BgmAudioSource.cs
@@ -15,11 +15,6 @@ public class BgmAudioSource: MonoBehaviour
     /// </summary>
     void Update()
     {
-        // 音量をセット
-        audioSource.volume = AudioParameter.bgmVolume;
-        // ミュートフラグをセット
-        audioSource.mute = AudioParameter.bgmMute;
-
         // 再生が終了したBGMはオブジェクトを切る
         if (!audioSource.isPlaying)
         {
